Use short for the task_get_priority function.

+++ src/bask_task.h
@@ -35,7 +35,7 @@ struct bask_task
 
 int task_check_input_nbrs (int, short, short, short);
 int task_check_input (bask_core*, char*, char*, char*, char*, char*, short);
-int task_get_priority (char*);
+short task_get_priority (char*);
 
 void task_free_ll (struct bask_task**);
 int task_insert (struct bask_task**, unsigned int, unsigned int, short, short, int, char*, char*, char*, char*, char*);
